Larger Refactor
